[refactor] 이미지 선택 영역을 ChangeProfileImageSelector 컴포넌트로 분리
